refactor: server & app separated
