+++ package.json
@@ -2,15 +2,15 @@
   "name": "code-treats",
   "version": "1.0.0",
   "description": "Interactive platform for running, saving and sharing code snippets.",
-  "main": "server/index.ts",
+  "main": "server/server.ts",
   "engines": {
     "node": "12.18.3",
     "npm": "6.14.8",
     "yarn": "1.22.4"
   },
   "scripts": {
-    "start": "cross-env NODE_ENV=production node prod/index.js",
-    "server": "nodemon server/index.ts",
+    "start": "cross-env NODE_ENV=production node prod/server.js",
+    "server": "nodemon server/server.ts",
     "client": "npm --prefix ./client run start",
     "client:install": "npm --prefix ./client install",
     "server:install": "npm install",
